# core/docs/swagger.yaml
definitions:
  define.AdminChangeUser:
    properties:
      forbid:
        description: '禁止用户: 1 未禁止 2 禁止登陆'
        type: boolean
      id:
        description: user id
        type: string
      password:
        description: 用户密码 Common
        type: string
      remark:
        description: 备注 Common
        type: string
      role:
        description: '用户类型: 1 普通用户 2 管理员'
        type: integer
    required:
    - id
    - role
    type: object
...
  define.ChangeUserSelf:
    properties:
      dingphone:
        description: dingding phone
        type: string
      email:
        description: 用户邮箱
        type: string
      id:
        description: user id
        type: string
      password:
        type: string
      remark:
        type: string
      telegram:
        description: telegram bot chat id
        type: string
      wechat:
        description: wechat id
        type: string
    required:
    - id
    type: object
...
  define.RegistryUser:
    properties:
      name:
        description: 用户名
        type: string
      password:
        description: 用户密码
        type: string
      remark:
        description: 备注
        type: string
      role:
        description: '用户类型: 1 普通用户 2 管理员'
        type: integer
    required:
    - name
    - password
    - role
    type: object
  resp.Response:
    properties:
      code:
        description: msg
        type: integer
      count:
        description: data count
        type: integer
      data:
        description: data
        type: object
      msg:
        description: code
        type: string
    type: object
info:
  contact:
    email: labulakalia@gmail.com
    name: labulaka521
    url: http://www.swagger.io/support
  description: Crocodile Swaager JSON API
  license:
    name: MIT 2.0
    url: https://crocodile/blob/master/LICENSE
  termsOfService: https://crocodile
  title: Crocidle API
  version: "1.0"
paths:
...
  /api/v1/user/admin:
    put:
      description: admin change user's role,forbid,password,remark
      parameters:
      - description: Admin Change User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/define.AdminChangeUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: admin change user info
      tags:
      - User
  /api/v1/user/all:
    get:
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get all users info
      tags:
      - User
  /api/v1/user/info:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get user info by token
      tags:
      - User
    put:
      description: change self config,like email,wechat,dingphone,slack,telegram,password,remark
      parameters:
      - description: Change Self User Info
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/define.ChangeUserSelf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: user change self's config info
      tags:
      - User
  /api/v1/user/login:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BasicAuth: []
      summary: login user
      tags:
      - User
  /api/v1/user/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BasicAuth: []
      summary: logout user
      tags:
      - User
  /api/v1/user/registry:
    post:
      parameters:
      - description: registry user
        in: body
        name: Registry
        required: true
        schema:
          $ref: '#/definitions/define.RegistryUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: registry new user
      tags:
      - User
  /api/v1/user/select:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BasicAuth: []
      summary: return name,id
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"